fix comment thai

+++ app/Http/Controllers/Api/OrderController.php
@@ -120,36 +120,39 @@ class OrderController extends ApiController
     */
     public function update(Order $order, CreateOrderRequest $request)
     {
-        try {
-            $total = 0;
-            $products = [];
-            DB::beginTransaction();
-            if ($request->product) {
-                foreach ($request->product as $product) {
-                    $orderDetail = OrderDetail::where('order_id', $order->id)->where('product_id', $product['id'])->first();
-                    if ($orderDetail) {
-                        $orderDetail->update([
-                            'quantity' => $product['quantity'],
-                        ]);
-                        $total += $product['quantity'] * $orderDetail->price;
-                        array_push($products, $product);
+        if ($order->status == Order::PENDING) {
+            try {
+                $total = 0;
+                $products = [];
+                DB::beginTransaction();
+                if ($request->product) {
+                    foreach ($request->product as $product) {
+                        $orderDetail = OrderDetail::where('order_id', $order->id)->where('product_id', $product['id'])->first();
+                        if ($orderDetail) {
+                            $orderDetail->update([
+                                'quantity' => $product['quantity'],
+                            ]);
+                            $total += $product['quantity'] * $orderDetail->price;
+                            array_push($products, $product);
+                        }
                     }
+                    OrderDetail::where('order_id', $order->id)->whereNotIn('product_id', array_pluck($request->product, 'id'))->delete();
+                } else {
+                    $order->update([
+                        "status" => Order::REJECTED,
+                    ]);
                 }
-                OrderDetail::where('order_id', $order->id)->whereNotIn('product_id', array_pluck($request->product, 'id'))->delete();
-            } else {
                 $order->update([
-                    "status" => Order::REJECTED,
+                    "address" => $request->address,
+                    "total" => $total,
                 ]);
+                DB::commit();
+                return $this->showOne($order->load('orderDetails'), Response::HTTP_OK);
+            } catch (Exception $e) {
+                DB::rollBack();
+                return $this->errorResponse(trans('errors.update_fail'), Response::HTTP_UNPROCESSABLE_ENTITY);
             }
-            $order->update([
-                "address" => $request->address,
-                "total" => $total,
-            ]);
-            DB::commit();
-            return $this->showOne($order->load('orderDetails'), Response::HTTP_OK);
-        } catch (Exception $e) {
-            DB::rollBack();
-            return $this->errorResponse(trans('errors.update_fail'), Response::HTTP_UNPROCESSABLE_ENTITY);
         }
+        return $this->errorResponse(trans('errors.update_fail'), Response::HTTP_UNPROCESSABLE_ENTITY);        
     }
 }
